fix 401

+++ src/posts/posts.service.ts
@@ -2,7 +2,7 @@
 import { Injectable } from '@nestjs/common';
 import { Posts } from './posts.model';
 import { InjectModel } from '@nestjs/mongoose';
-import { Conflict, NotFound, BadRequest } from 'http-errors';
+import { Conflict, NotFound, BadRequest, Unauthorized } from 'http-errors';
 import { CreatePostDto } from './dto/create.post.dto';
 import { User } from 'src/users/users.model';
 import { VerifyPostDto } from './dto/verify.post.dto';
@@ -17,8 +17,11 @@ export class PostsService {
   ) {}
 
   async findAllPosts(req: any) {
+    const user = await this.userService.findToken(req);
+    if (!user) {
+      throw new Unauthorized('jwt expired');
+    }
     try {
-      const user = await this.userService.findToken(req);
       if (user.role === 'admin' || user.role === 'moderator') {
         return await this.postModel.find().exec();
       }
@@ -28,8 +31,11 @@ export class PostsService {
   }
 
   async findNewPosts(req: any) {
+    const user = await this.userService.findToken(req);
+    if (!user) {
+      throw new Unauthorized('jwt expired');
+    }
     try {
-      const user = await this.userService.findToken(req);
       if (user.role === 'admin' || user.role === 'moderator') {
         return await this.postModel.find({ verify: 'new' }).exec();
       }
@@ -39,11 +45,13 @@ export class PostsService {
   }
 
   async findMyPosts(req: any) {
-    try {      
-      const user = await this.userService.findToken(req);
+    const user = await this.userService.findToken(req);
+    if (!user) {
+      throw new Unauthorized('jwt expired');
+    }
+    try {
       if (user) {
-        
-        const post = await this.postModel.find({ "owner.id": user.id });
+        const post = await this.postModel.find({ 'owner.id': user.id });
         return post;
       }
     } catch (e) {
@@ -70,8 +78,11 @@ export class PostsService {
   }
 
   async createPost(post: CreatePostDto, req: any): Promise<Posts> {
+    const user = await this.userService.findToken(req);
+    if (!user) {
+      throw new Unauthorized('jwt expired');
+    }
     try {
-      const user = await this.userService.findToken(req);
       if (user) {
         const createdPost = await this.postModel.create(post);
         createdPost.save();
@@ -92,8 +103,11 @@ export class PostsService {
   }
 
   async updatePost(post: CreatePostDto, id: string, req: any): Promise<Posts> {
+    const user = await this.userService.findToken(req);
+    if (!user) {
+      throw new Unauthorized('jwt expired');
+    }
     try {
-      const user = await this.userService.findToken(req);
       if (user) {
         const { ...params } = post;
         await this.postModel.findByIdAndUpdate({ _id: id }, { ...params });
@@ -106,9 +120,13 @@ export class PostsService {
   }
 
   async deletePost(id: string, req: any): Promise<Posts> {
+    const user = await this.userService.findToken(req);
+    const post = await this.postModel.findById({ _id: id });
+    if (!user) {
+      throw new Unauthorized('jwt expired');
+    }
+
     try {
-      const user = await this.userService.findToken(req);
-      const post = await this.postModel.findById({ _id: id });
       if (user.role === 'admin' || user.role === 'moderator') {
         const find = await this.postModel.findByIdAndRemove({ _id: id }).exec();
         return find;
@@ -128,16 +146,20 @@ export class PostsService {
     req: any,
     postUp: VerifyPostDto,
   ): Promise<Posts> {
+    const admin = await this.userService.findToken(req);
+    const post = await this.postModel.findById(id);
+    if (!admin) {
+      throw new Unauthorized('jwt expired');
+    }
     try {
-      const admin = await this.userService.findToken(req);
-      const post = await this.postModel.findById(id);
-
       if (!admin || !post) {
         throw new Conflict('Not found');
       }
 
-      if (admin.role === 'admin' || admin.role === 'moderator' && post.verify === 'new')
-      {
+      if (
+        admin.role === 'admin' ||
+        (admin.role === 'moderator' && post.verify === 'new')
+      ) {
         const { ...params } = postUp;
         await this.postModel.findByIdAndUpdate({ _id: id }, { ...params });
         post.save();
@@ -151,10 +173,12 @@ export class PostsService {
   }
 
   async favoritePost(id: string, req: any): Promise<Posts> {
+    const user = await this.userService.findToken(req);
+    const post = await this.postModel.findById(id);
+    if (!user) {
+      throw new Unauthorized('jwt expired');
+    }
     try {
-      const user = await this.userService.findToken(req);
-      const post = await this.postModel.findById(id);
-
       if (!user || !post) {
         throw new Conflict('Not found');
       }
@@ -202,8 +226,11 @@ export class PostsService {
   }
 
   async findMyFavPosts(req: any) {
+    const user = await this.userService.findToken(req);
+    if (!user) {
+      throw new Unauthorized('jwt expired');
+    }
     try {
-      const user = await this.userService.findToken(req);
       const post = await this.postModel.find({ favorite: user.id }).exec();
       return post;
     } catch (e) {
